Add method name to TCD trace definition
Currently for TCD traces the method can only hold an external reference
but this is not very user-friendly and often it is more convent to show
the method name instead.

This adds an (optional) method name to the method description as it is
already the case for the FID trace.